[FIX] Fix tests and actions (#6)
- Replace actions with actions from repo e-budget-actions
- Add Flyway dependency
- Add Lombok dependency
- Move from Double to BigDecimal data types due to precision accuracy
- Add test profile with database devservice configurations to run tests
- Refactor/Fix E2E and Unit tests

+++ e-budget-be/src/main/java/com/ebudget/account/model/Account.java
@@ -15,6 +15,7 @@ import lombok.NoArgsConstructor;
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
@@ -36,8 +37,8 @@ public class Account {
     @Column(name = "account_type")
     private AccountType accountType;
     @Column(name = "initial_balance")
-    private Double initialBalance;
-    private Double balance;
+    private BigDecimal initialBalance;
+    private BigDecimal balance;
     @CreationTimestamp
     @Column(name = "created_at")
     private LocalDateTime createdAt;
